requirements.txt
django==2.2.26
djangorestframework==3.12.2
markdown
django-filter
ipython
autopep8
pylint
